fix: resolve websocket service conflict
