+++ client/src/App.tsx
@@ -1,12 +1,12 @@
 import { QueryClientProvider } from '@tanstack/react-query';
-import { useEffect } from 'react';
 import { Switch, Route } from 'wouter';
 
 import { queryClient } from './lib/queryClient';
 
 import { Toaster } from '@/components/ui/toaster';
 import { TooltipProvider } from '@/components/ui/tooltip';
-import { initTelegram } from '@/lib/telegram';
+import { TelegramProvider } from '@/services/telegram';
+import { WebSocketProvider } from '@/services/websocket';
 import AdminPage from '@/pages/admin';
 import NotFound from '@/pages/not-found';
 import TournamentDetailPage from '@/pages/tournament-detail';
@@ -24,17 +24,17 @@ function Router() {
 }
 
 function App() {
-  useEffect(() => {
-    initTelegram();
-  }, []);
-
   return (
     <QueryClientProvider client={queryClient}>
       <TooltipProvider>
-        <div className="min-h-screen bg-gray-100">
-          <Toaster />
-          <Router />
-        </div>
+        <TelegramProvider>
+          <WebSocketProvider>
+            <div className="min-h-screen bg-gray-100">
+              <Toaster />
+              <Router />
+            </div>
+          </WebSocketProvider>
+        </TelegramProvider>
       </TooltipProvider>
     </QueryClientProvider>
   );
